Edited the protocol to create a lamella of correct shape.
Removed the last remaining references to 'window' in functions outside of the window class.

Removed a host of bugs to do with the UI, such as pressing buttons that required an experiment to be created/loaded, or the protocol to be loaded where none was.

Fixed bugs related to clicking 'create expriment', 'load experiment', 'load protocol' and then not following the prompts.

Fixed bugs related to the updating of 'lamella ready' in the UI after milling fiducials

Fixed bugs related to the milling process of trenches. The function was returning after milling a single trench.